requirements.txt
django-oauth-toolkit
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
Django==2.2.7
django-filter==2.2.0
django-pandas==0.6.1
djangorestframework==3.10.3
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mysqlclient==1.4.6
numpy==1.17.4
pandas==0.25.3
patsy==0.5.1
Pillow==6.2.1
Pygments==2.5.2
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
scipy==1.3.3
six==1.13.0
sqlparse==0.3.0
statsmodels==0.10.2
uritemplate==3.0.0
urllib3==1.25.7
